Move join validator from valid_join.py to join_validator.py
The original name conflicts a little bit with our existing
conventions. Moving this now before adding new tests or making
changes to the file.
